Interactables! Read description...
Added in basic cubes that can be approached and triggered by pressing
the Spacebar. A closeup is then shown on screen, hit Spacebar again to
back out and continue as normal.

You can add more Interactables to the scene by dragging up the PreFab
(of the same name), then drag the Player object to the inspector under
the Interaction Script. For some reason you can't assign objects to a
PreFab by default :/