Room Tracking and Fading Walls
Added in some scripts to keep track of the player's current room. Each
Floor object has a trigger collider that sets the active room. The walls
also have identifiers to help tell them when to fade out and such.

Pay attention to how the scripts are set up in the inspector at current
if you want to add more floors/walls etc.